Move kafka serdes to a separate directory and remove deprecated classes
