--- serdes/avro-serde/pom.xml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-  <parent>
-    <groupId>io.apicurio</groupId>
-    <artifactId>apicurio-registry</artifactId>
-    <version>3.0.0-SNAPSHOT</version>
-    <relativePath>../../pom.xml</relativePath>
-  </parent>
-
-  <artifactId>apicurio-registry-serdes-avro-serde</artifactId>
-  <packaging>jar</packaging>
-  <name>apicurio-registry-serdes-avro-serde</name>
-
-  <properties>
-    <projectRoot>${project.basedir}/../..</projectRoot>
-  </properties>
-
-  <dependencies>
-    <dependency>
-      <groupId>io.apicurio</groupId>
-      <artifactId>apicurio-registry-serde-common</artifactId>
-    </dependency>
-
-    <dependency>
-      <groupId>org.apache.avro</groupId>
-      <artifactId>avro</artifactId>
-    </dependency>
-  </dependencies>
-
-</project>
