+++ serdes/kafka/avro-serde/src/main/java/io/apicurio/registry/serde/avro/strategy/RecordIdStrategy.java
