+++ serdes/kafka/protobuf-serde/src/main/proto/ref.proto
