--- serdes/serde-common/src/main/java/io/apicurio/registry/serde/SchemaParser.java
@@ -1,14 +1,0 @@
-package io.apicurio.registry.serde;
-
-/**
- * This class is deprecated and eventually will be replaced by
- * {@link io.apicurio.registry.resolver.SchemaParser}
- */
-@Deprecated
-public interface SchemaParser<S> {
-
-    public String artifactType();
-
-    public S parseSchema(byte[] rawSchema);
-
-}
